Duplicate DocIdSet data into the termIdMap for shorter lookups
For a TermId -> DocIdSet lookup we previously had to do two lookups,
one in the TermId -> Term mapping and then in the Term -> DocIdSet
mapping. The auto-complete stuff relies quite heavily on that lookup
and doing one lookup rather than two nearly doubles  performance for
the auto-complete.

The space cost is minimal since it's the same data, just accessable
via two indexes.